Revert "[internal] Consolidate PyO3 native extensions (#13614)"
This reverts commit ccd9b836b877bbd3b49bb50a0f558e6734b01972.
