+++ .github/workflows/test-cron.yaml
@@ -100,6 +100,8 @@ jobs:
 
           src/python/pants/engine/internals/native_engine.so
 
+          src/python/pants/engine/internals/native_engine_pyo3.so
+
           src/python/pants/engine/internals/native_engine.so.metadata'
     - if: github.event_name != 'pull_request'
       name: Setup toolchain auth
@@ -139,6 +141,8 @@ jobs:
         path: '.pants
 
           src/python/pants/engine/internals/native_engine.so
+
+          src/python/pants/engine/internals/native_engine_pyo3.so
 
           src/python/pants/engine/internals/native_engine.so.metadata'
     - if: '!contains(env.COMMIT_MESSAGE, ''[ci skip-rust]'')'
@@ -251,6 +255,8 @@ jobs:
 
           src/python/pants/engine/internals/native_engine.so
 
+          src/python/pants/engine/internals/native_engine_pyo3.so
+
           src/python/pants/engine/internals/native_engine.so.metadata'
     - if: github.event_name != 'pull_request'
       name: Setup toolchain auth
@@ -268,6 +274,8 @@ jobs:
         path: '.pants
 
           src/python/pants/engine/internals/native_engine.so
+
+          src/python/pants/engine/internals/native_engine_pyo3.so
 
           src/python/pants/engine/internals/native_engine.so.metadata'
     - env:
