+++ build-support/bin/generate_github_workflows.py
@@ -35,6 +35,7 @@ Env = Dict[str, str]
 NATIVE_FILES = [
     ".pants",
     "src/python/pants/engine/internals/native_engine.so",
+    "src/python/pants/engine/internals/native_engine_pyo3.so",
     "src/python/pants/engine/internals/native_engine.so.metadata",
 ]
