+++ src/rust/engine/Cargo.toml
@@ -26,6 +26,7 @@ members = [
   "cache",
   "client",
   "concrete_time",
+  "engine_pyo3",
   "fs",
   "fs/brfs",
   "fs/fs_util",
@@ -67,6 +68,7 @@ default-members = [
   "cache",
   "client",
   "concrete_time",
+  "engine_pyo3",
   "fs",
   "fs/fs_util",
   "fs/store",
